# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Digital Service BKKBN for Integration - Wilayah Service",
        "title": "BKKBN Digital Service - Wilayah",
        "termsOfService": "https://bkkbn.go.id",
        "contact": {
            "name": "Direktorat Teknologi Informasi dan Data",
            "url": "https://bkkbn.go.id",
            "email": "dittifdok@bkkbn.go.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/api/2022/wilayah",
    "paths": {
        "/showkabupaten": {
            "get": {
                "description": "Get Wilayah By Kabupaten",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Wilayah By Kabupaten"
                ],
                "summary": "Get Wilayah By Kabupaten",
                "operationId": "get-list-wilayah-by-kabupaten",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province Code",
                        "name": "kdprov",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of Kabupaten per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kabupaten.Kabupaten"
                        }
                    }
                }
            }
        },
        "/showkecamatan": {
            "get": {
                "description": "Get Wilayah By Kecamatan",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Wilayah By Kecamatan"
                ],
                "summary": "Get Wilayah By Kecamatan",
                "operationId": "get-list-wilayah-by-kecamatan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province Code",
                        "name": "kdprov",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kabupaten Code",
                        "name": "kdkab",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of Kecamatan per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kecamatan.Kecamatan"
                        }
                    }
                }
            }
        },
        "/showkelurahan": {
            "get": {
                "description": "Get Wilayah Sasaran By Kelurahan",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Wilayah By Kelurahan"
                ],
                "summary": "Get Wilayah By Kelurahan",
                "operationId": "get-list-wilayah-by-kelurahan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Province Code",
                        "name": "kdprov",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kabupaten Code",
                        "name": "kdkab",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kecamatan Code",
                        "name": "kdkec",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of Kelurahan per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kelurahan.Kelurahan"
                        }
                    }
                }
            }
        },
        "/showprovinsi": {
            "get": {
                "description": "Get Wilayah by Province",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Wilayah by Province"
                ],
                "summary": "Get Wilayah by Province",
                "operationId": "get-wilayah-by-province",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 34,
                        "description": "Number of Province per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/provinsi.Provinsi"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "kabupaten.Kabupaten": {
            "type": "object",
            "properties": {
                "idKabupaten": {
                    "type": "integer"
                },
                "idProvinsi": {
                    "type": "integer"
                },
                "kodeKabupaten": {
                    "type": "integer"
                },
                "kodeProvinsi": {
                    "type": "integer"
                },
                "namaKabupaten": {
                    "type": "string"
                },
                "namaProvinsi": {
                    "type": "string"
                }
            }
        },
        "kecamatan.Kecamatan": {
            "type": "object",
            "properties": {
                "idKabupaten": {
                    "type": "integer"
                },
                "idKecamatan": {
                    "type": "integer"
                },
                "idProvinsi": {
                    "type": "integer"
                },
                "kodeKabupaten": {
                    "type": "integer"
                },
                "kodeKecamatan": {
                    "type": "integer"
                },
                "kodeProvinsi": {
                    "type": "integer"
                },
                "namaKabupaten": {
                    "type": "string"
                },
                "namaKecamatan": {
                    "type": "string"
                },
                "namaProvinsi": {
                    "type": "string"
                }
            }
        },
        "kelurahan.Kelurahan": {
            "type": "object",
            "properties": {
                "idKabupaten": {
                    "type": "integer"
                },
                "idKecamatan": {
                    "type": "integer"
                },
                "idKelurahan": {
                    "type": "integer"
                },
                "idProvinsi": {
                    "type": "integer"
                },
                "kodeKabupaten": {
                    "type": "integer"
                },
                "kodeKecamatan": {
                    "type": "integer"
                },
                "kodeKelurahan": {
                    "type": "integer"
                },
                "kodeProvinsi": {
                    "type": "integer"
                },
                "namaKabupaten": {
                    "type": "string"
                },
                "namaKecamatan": {
                    "type": "string"
                },
                "namaKelurahan": {
                    "type": "string"
                },
                "namaProvinsi": {
                    "type": "string"
                }
            }
        },
        "provinsi.Provinsi": {
            "type": "object",
            "properties": {
                "idProvinsi": {
                    "type": "integer"
                },
                "kodeProvinsi": {
                    "type": "integer"
                },
                "namaProvinsi": {
                    "type": "string"
                }
            }
        }
    }
}